permutations, tqdm, pickle results

+++ Pairwise_Temporal_ISC_vs_consensus.py
@@ -13,6 +13,7 @@ import tqdm
 import numpy as np
 import pandas as pd
 from scipy.stats import pearsonr
+import matplotlib.pyplot as plt
 
 from ISC_Helper import compute_isc, get_rois
 
@@ -117,31 +118,40 @@ df_consensus = np.mean(rolling_mean, axis=1)
 #             print(f"{roi} and {emotions[emotion]}: {corrs_roi[roi].loc[emotions[emotion]]['r']:.3f}, "
 #                   f"{corrs_roi[roi].loc[emotions[emotion]]['p']:.3f}")
 
-# do the correlation voxelwise, ISC vs consensus
-start = time.time()
-s_map = np.empty(shape=(iscs_roi_selected['wholebrain'].shape[1], 2))
-x = iscs_roi_selected['wholebrain']
-print('computing s_map')
-for i in range(x.shape[0]):
-    s_map[i] = pearsonr(x.T[i], df_consensus[:, 2])
-    
-# save s_map to pickle
-with open(f"{data_path}/s_map.pkl", 'wb') as f:
-    pickle.dump(s_map, f)
 
-start = time.time()
-rng = np.random.default_rng()
 n_perm = 10000
-perm = np.empty(shape=(n_perm, x.shape[1], 2))
-for i in tqdm(range(n_perm)):
-    # if i%1000==0 or i==100:
-    #     print(f"permutation {i}")
-    for j in range(x.shape[0]):
-        perm[i, j] = pearsonr(x.T[j], rng.permutation(df_consensus[:, 2]))
-end = time.time()
-print(end - start)
+if not os.path.exists(f"{data_path}/perm{n_perm}.pkl"):
+    # do the correlation voxelwise, ISC vs consensus
+    start = time.time()
+    s_map = np.empty(shape=(iscs_roi_selected['wholebrain'].shape[1], 2))
+    x = iscs_roi_selected['wholebrain']
+    print('computing s_map')
+    for i in range(x.shape[0]):
+        s_map[i] = pearsonr(x.T[i], df_consensus[:, 2])
+        
+    # save s_map to pickle
+    with open(f"{data_path}/s_map.pkl", 'wb') as f:
+        pickle.dump(s_map, f)
 
-# save perm to pickle
-with open(f"{data_path}/perm{n_perm}.pkl", 'wb') as f:
-    pickle.dump(perm, f)
+    start = time.time()
+    rng = np.random.default_rng()
+    perm = np.empty(shape=(n_perm, x.shape[1], 2))
+    for i in tqdm(range(n_perm)):
+        # if i%1000==0 or i==100:
+        #     print(f"permutation {i}")
+        for j in range(x.shape[0]):
+            perm[i, j] = pearsonr(x.T[j], rng.permutation(df_consensus[:, 2]))
+    end = time.time()
+    print(end - start)
 
+    # save perm to pickle
+    with open(f"{data_path}/perm{n_perm}.pkl", 'wb') as f:
+        pickle.dump(perm, f)
+
+else:
+    with open(f"{data_path}/perm{n_perm}.pkl", 'rb') as f:
+        perm = pickle.load(f)
+
+# view histogram
+plt.hist(perm[:, 0, 0], bins=100)
+plt.show()
